+++ Helping.cs
@@ -1,12 +1,12 @@
 namespace SystemOfEquations;
 
-internal record Helping(Food Food, double Servings)
+internal record Helping(Food Food, decimal Servings)
 {
     public override string ToString() => Food.ToString(Servings);
 
     public Macros Macros => Food.NutritionalInformation.Macros * Servings;
     public NutritionalInformation NutritionalInformation => Food.NutritionalInformation * Servings;
-    public static Helping operator *(Helping h, double d) => new(h.Food, h.Servings * d);
+    public static Helping operator *(Helping h, decimal d) => new(h.Food, h.Servings * d);
 }
 
 internal static class HelpingExtensions
